requirements.txt
mysql-connector-python==9.3.0
flask==2.3.2
flask-cors==3.0.10